feat: moving getValueFromLogs and coinToNumber in utils package
