+++ packages/cosmos-utils/package.json
@@ -21,7 +21,8 @@
   "dependencies": {
     "@babel/runtime": "^7.12.1",
     "@liquality/types": "^1.4.3",
-    "@cosmjs/proto-signing": "^0.25.5",
+    "@liquality/cosmos-networks": "^1.4.3",
+    "@cosmjs/crypto": "^0.25.5",
     "@cosmjs/encoding": "^0.25.5",
     "@cosmjs/stargate": "^0.25.5"
   },
